% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_NOAA}
\alias{clean_NOAA}
\title{The NOAA Significant Earthquake Database}
\format{A data frame with 5901 rows and 48 variables:
\describe{
 \item{YEAR}{Year (Century and Year) Valid values: -2150 to Present.  Format +/-yyyy (-is B.C, +is A.D.) The Date and Time are given in Universal Coordinated Time (also known as Greenwich Mean Time). The local date may be one day different. }
 \item{MONTH}{Month.  Valid values: 1-12 The Date and Time are given in Universal Coordinated Time (also known as Greenwich Mean Time). The local date may be one day different.}
 \item{DAY}{Day.  Valid values: 1-31 (where months apply)  The Date and Time are given in Universal Coordinated Time (also known as Greenwich Mean Time). The local date may be one day different.}
 \item{HOUR}{Hour.  Valid values: 0-23  The Date and Time are given in Universal Coordinated Time (also known as Greenwich Mean Time). The local date may be one day different.}
 \item{MINUTE}{Minute.  Valid values: 0-59  The Date and Time are given in Universal Coordinated Time (also known as Greenwich Mean Time). The local date may be one day different.}
 \item{SECOND}{Second.  Valid values: 0-59  The Date and Time are given in Universal Coordinated Time (also known as Greenwich Mean Time). The local date may be one day different.}
 \item{DATE}{YEAR, MONTH, and DAY values combined into a valid date object}
 \item{FOCAL_DEPTH}{Focal depth (km):  Valid values: 0 to 700 km  The depth of the earthquake is given in kilometers.}
 \item{EQ_PRIMARY}{The primary magnatude measure for the earthquake.  See EQ_MAG_ML and EQ_MAG_MFA below for more information}
 \item{EQ_MAG_MW}{Mw Magnitude: Valid values: 0.0 to 9.9  The Mw magnitude is based on the moment magnitude scale. Moment is a physical quantity proportional to the slip on the fault times the area of the fault surface that slips; it is related to the total energy released in the EQ. The moment can be estimated from seismograms (and also from geodetic measurements). The moment is then converted into a number similar to other earthquake magnitudes by a standard formula. The result is called the moment magnitude. The moment magnitude provides an estimate of earthquake size that is valid over the complete range of magnitudes, a characteristic that was lacking in other magnitude scales.  The magnitude is a measure of seismic energy. The magnitude scale is logarithmic. An increase of one in magnitude represents a tenfold increase in the recorded wave amplitude. However, the energy release associated with an increase of one in magnitude is not tenfold, but about thirtyfold. For example, approximately 900 times more energy is released in an earthquake of magnitude 7 than in an earthquake of magnitude 5. Each increase in magnitude of one unit is equivalent to an increase of seismic energy of about 1.6 x 10,000,000,000,000 ergs.}
 \item{EQ_MAG_MS}{Ms Magnitude: Valid values: 0.0 to 9.9 The Ms magnitude is the surface-wave magnitude of the earthquake. The magnitude is a measure of seismic energy. The magnitude scale is logarithmic. An increase of one in magnitude represents a tenfold increase in the recorded wave amplitude. However, the energy release associated with an increase of one in magnitude is not tenfold, but about thirtyfold. For example, approximately 900 times more energy is released in an earthquake of magnitude 7 than in an earthquake of magnitude 5. Each increase in magnitude of one unit is equivalent to an increase of seismic energy of about 1.6 x 10,000,000,000,000 ergs.}
 \item{EQ_MAG_MB}{mb Magnitude: Valid values: 0.0 to 9.9 The Mb magnitude is the compressional body wave (P-wave) magnitude. The magnitude is a measure of seismic energy. The magnitude scale is logarithmic. An increase of one in magnitude represents a tenfold increase in the recorded wave amplitude. However, the energy release associated with an increase of one in magnitude is not tenfold, but about thirtyfold. For example, approximately 900 times more energy is released in an earthquake of magnitude 7 than in an earthquake of magnitude 5. Each increase in magnitude of one unit is equivalent to an increase of seismic energy of about 1.6 x 10,000,000,000,000 ergs.}
 \item{EQ_MAG_ML}{ML Magnitude: Valid values: 0.0 to 9.9 The ML magnitude was the original magnitude relationship defined by Richter and Gutenberg for local earthquakes in 1935. It is based on the maximum amplitude of a seismogram recorded on a Wood-Anderson torsion seismograph. Although these instruments are no longer widely in use, ML values are calculated using modern instrumentation with appropriate adjustments. The magnitude is a measure of seismic energy. The magnitude scale is logarithmic. An increase of one in magnitude represents a tenfold increase in the recorded wave amplitude. However, the energy release associated with an increase of one in magnitude is not tenfold, but about thirtyfold. For example, approximately 900 times more energy is released in an earthquake of magnitude 7 than in an earthquake of magnitude 5. Each increase in magnitude of one unit is equivalent to an increase of seismic energy of about 1.6 x 10,000,000,000,000 ergs.}
 \item{EQ_MAG_MFA}{Mfa Magnitude: Valid values: 0.0 to 9.9 The Mfa magnitudes are computed from the felt area, for earthquakes that occurred before seismic instruments were in general use. The magnitude is a measure of seismic energy. The magnitude scale is logarithmic. An increase of one in magnitude represents a tenfold increase in the recorded wave amplitude. However, the energy release associated with an increase of one in magnitude is not tenfold, but about thirtyfold. For example, approximately 900 times more energy is released in an earthquake of magnitude 7 than in an earthquake of magnitude 5. Each increase in magnitude of one unit is equivalent to an increase of seismic energy of about 1.6 x 10,000,000,000,000 ergs.}
 \item{EQ_MAG_UNK}{Unknown Magnitude: Valid values: 0.0 to 9.9 The computational method for the earthquake magnitude was unknown and could not be determined from the published sources. The magnitude is a measure of seismic energy. The magnitude scale is logarithmic. An increase of one in magnitude represents a tenfold increase in the recorded wave amplitude. However, the energy release associated with an increase of one in magnitude is not tenfold, but about thirtyfold. For example, approximately 900 times more energy is released in an earthquake of magnitude 7 than in an earthquake of magnitude 5. Each increase in magnitude of one unit is equivalent to an increase of seismic energy of about 1.6 x 10,000,000,000,000 ergs.}
 \item{INTENSITY}{MMI Intensity: Valid values: 1 to 12 The Modified Mercalli Intensity (Int) is given in Roman Numerals (converted to numbers in the digital database).}
 \item{COUNTRY}{The Country where the Earthquake occurred}
 \item{STATE}{The two-letter State or Province abbreviation where the Earthquake occurred}
 \item{LOCATION_NAME}{The Country, State, Province or City where the Earthquake occurred (For example enter: USA or California or San Francisco).  This is only an approximate geographic location. Events prior to 1900 were not instrumentally located, therefore, the location given is based on the latitude and longitude of the city where the damage occurred. In the Geographic Location column, the country is listed first, then the province or state, and finally the city or cities where damage was reported. If there are different spellings of a city name the additional spellings are in parentheses.}
 \item{LATITUDE}{ Latitude: 0 to 90 (Northern Hemisphere) -90 to 0 (Southern Hemisphere) Events prior to 1900 were not instrumentally located, therefore, the location given is based on the latitude and longitude of the city where the damage occurred. }
 \item{LONGITUDE}{ Longitude: 0 to 180 (Eastern Hemisphere) -180 to 0 (Western Hemisphere) Events prior to 1900 were not instrumentally located, therefore, the location given is based on the latitude and longitude of the city where the damage occurred. }
 \item{REGION_CODE}{Region Code: Regional boundaries defined as follows:     10 = 	Central, Western and S. Africa, 15 = 	Northern Africa, 20 = 	Antarctica
       30 = 	East Asia, 40 = 	Central Asia and Caucasus, 50 = 	Kamchatka and Kuril Islands, 60 = 	S. and SE. Asia and Indian Ocean,
       70 = 	Atlantic Ocean, 80 = 	Bering Sea, 90 = 	Caribbean,  100 = 	Central America, 110 = 	Eastern Europe, 120 = 	Northern and Western Europe,
       130 = 	Southern Europe, 140 = 	Middle East, 150 = 	North America and Hawaii, 160 = 	South America, 170 = 	Central and South Pacific}
 \item{FLAG_TSUNAMI}{When a tsunami was generated by an earthquake, "Tsu" appears}
 \item{DEATHS}{Number of Deaths from the Earthquake: Valid values: 0 to 1100000 Whenever possible, numbers of deaths are listed. When references give different death tolls, each death toll and corresponding reference is listed as a separate record.}
 \item{DEATHS_DESCRIPTION}{Valid values: 0 to 4 When a description was found in the historical literature instead of an actual number of deaths, this value was coded and listed in the Deaths De column. If the actual number of deaths was listed, a descriptor was also added for search purposes.
      0 = None, 1 = Few (~1 to 50 deaths), 2 = Some (~51 to 100 deaths), 3 = Many (~101 to 1000 deaths), 4 = Very Many (~1001 or more deaths) }
 \item{MISSING}{Number of Missing from the Earthquake: Valid values: 0 to 1100000 Whenever possible, numbers of missing are listed. When references give different numbers of missing, each missing number and corresponding reference is listed as a separate record.}
 \item{MISSING_DESCRIPTION}{ When a description was found in the historical literature instead of an actual number of missing, this value was coded and listed in the Missing De column. If the actual number of Missing was listed, a descriptor was also added for search purposes.
       0 = None, 1 = Few (~1 to 50 missing), 2 = Some(~51 to 100 missing), 3 = Many (~101 to 1000 missing), 4 = Very Many (~1001 or more missing) }
 \item{INJURIES}{Number of Injuries from the Earthquake: Valid values: 0 to 30000 Whenever possible, numbers of injuries from the earthquake are listed. }
 \item{INJURIES_DESCRIPTION}{Valid values: 0 to 4  When a description was found in the historical literature instead of an actual number of injuries, this value was coded and listed in the Injuries De column. If the actual number of injuries was listed, a descriptor was also added for search purposes.
       0 = None, 1 = Few (~1 to 50 injuries), 2 = Some(~51 to 100 injuries), 3 = Many (~101 to 1000 injuries), 4 = Very Many (~1001 or more injuries) }
 \item{DAMAGE_MILLIONS_DOLLARS}{Damage in Millions of Dollars from the Earthquake: The value in the Damage column should be multipled by 1,000,000 to obtain the actual dollar amount.
       When a dollar amount for damage was found in the literature, it was listed in the Damage column in millions of U.S. dollars. The dollar value listed is the value at the time of the event. To convert the damage to current dollar values, please use the Consumer Price Index Calculator. Monetary conversion tables for the time of the event were used to convert foreign currency to U.S. dollars. }
 \item{DAMAGE_DESCRIPTION}{Valid values: 0 to 4 For those events not offering a monetary evaluation of damage, the following five-level scale was used to classify damage (1990 dollars) and was listed in the Damage De column. If the actual dollar amount of damage was listed, a descriptor was also added for search purposes.
       0 = NONE, 1 = LIMITED (roughly corresponding to less than $1 million), 2 = MODERATE (~$1 to $5 million), 3 = SEVERE (~>$5 to $24 million), 4 = EXTREME (~$25 million or more)
       When possible, a rough estimate was made of the dollar amount of damage based upon the description provided, in order to choose the damage category. In many cases, only a single descriptive term was available. These terms were converted to the damage categories based upon the authors apparent use of the term elsewhere. In the absence of other information, LIMITED is considered synonymous with slight, minor, and light, SEVERE as synonymous with major, extensive, and heavy, and EXTREME as synonymous with catastrophic.
       Note: The descriptive terms relate approximately to current dollar values.}
 \item{HOUSES_DESTROYED}{Total Number of Houses Destroyed by the Earthquake: Valid values: 0 to 1100000 Whenever possible, numbers of houses destroyed are listed. }
 \item{HOUSES_DESTROYED_DESCRIPTION}{Valid values: 0 to 4  For those events not offering an exact number of houses destroyed, the following four-level scale was used to classify the destruction and was listed in the Houses Destroyed De column. If the actual number of houses destroyed was listed, a descriptor was also added for search purposes.
       0 = None, 1 = Few (~1 to 50 houses), 2 = Some (~51 to 100 houses), 3 = Many (~101 to 1000 houses), 4 = Very Many (~1001 or more houses) }
 \item{HOUSES_DAMAGED}{Total Number of Houses Damaged by the Earthquake: Valid values: 0 to 1100000  Whenever possible, numbers of houses damaged are listed. }
 \item{HOUSES_DAMAGED_DESCRIPTION}{Valid values: 0 to 4 For those events not offering an exact number of houses damaged, the following four-level scale was used to classify the damage and was listed in the Houses Damaged De column. If the actual number of houses destroyed was listed, a descriptor was also added for search purposes.
       0 = None, 1 = Few (~1 to 50 houses), 2 = Some (~51 to 100 houses), 3 = Many (~101 to 1000 houses), 4 = Very Many (~1001 or more houses) }
 \item{TOTAL_DEATHS}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_DEATHS_DESCRIPTION}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_MISSING}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_MISSING_DESCRIPTION}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_INJURIES}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_INJURIES_DESCRIPTION}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_DAMAGE_MILLIONS_DOLLARS}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_DAMAGE_DESCRIPTION}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_HOUSES_DESTROYED}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_HOUSES_DESTROYED_DESCRIPTION}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_HOUSES_DAMAGED}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{TOTAL_HOUSES_DAMAGED_DESCRIPTION}{Same as above, but adding Tsunami, Volcano, etc effects}
 \item{I_D}{The unique numeric identifier of the record}
}}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
clean_NOAA
}
\description{
A dataset containing location, strenght, and descriptive information for
5901 significant earthquake events worldwide since 2500 BCE
}
\keyword{datasets}
