% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_script.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Build an html label for earthquakes}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A dataframe containing location names in column \code{LOCATION_NAME},
earthquake magnatude in column\code{EQ_PRIMARY}, and total deaths caused in column
\code{TOTAL_DEATHS}.  The included EQ_NAA dataframe is already correctly formatted}
}
\value{
A character vector of formatted html strings to label the popups in
  \code{\link{eq_map}}
}
\description{
Create a nicely formatted HTML label for use with \code{\link{eq_map}} showing
Location, Magnatude, and Total Deaths.  Missing values are skipped in the created label
}
\examples{
eq_create_label(data.frame(
       LOCATION_NAME = "Nowhere",
       EQ_PRIMARY = 0.0, TOTAL_DEATHS = 0))


}
\seealso{
\code{\link{eq_map}}, \pkg{leaflet}
}
