% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{eq_good_date}
\alias{eq_good_date}
\title{Combine the year, month, and day columns in the NOAA earthquake data into a single date}
\usage{
eq_good_date(yr, mnth, dy)
}
\arguments{
\item{yr}{integer of the year the earthquake occurred.  Negative numbers indicate BCE}

\item{mnth}{integer of the month (1-12)}

\item{dy}{integer of the day of the month}
}
\value{
a single date object
}
\description{
Individual year, month, and day columns make filtering the data difficult.  This function
combines them into a single date object.  Only to be used internally to the package.
}
\details{
Note that R date objects do not handle BCE dates well.  For example, they assume
 a year 0 exists, when in fact dates go from 1 BCE to 1 CE, with no 0 in between.  Therefore,
 date differences between BCE and CE will not be correct.  However, date ranges should still be
 accurately reflected.

 Not all entries have date details beyond year, in that case the date is
 January 1 of the appropriate year
}
\examples{
eq_good_date(2017, 7, 1)

}
