% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean up location names in the NOAA earthquake database}
\usage{
eq_location_clean(loc_nm)
}
\arguments{
\item{loc_nm}{String containing original location name from the NOAA earthquake file}
}
\value{
A string removing the country name, fixing spacing, and converting
to Title Case
}
\description{
The LOCATION_NAME column in the NOAA database starts out life including
the country name, as well as other formatting issues.  Use this function
to strip out and reformat the region for convenient labeleing
}
\examples{
QuakeCoursera:::eq_location_clean("USA: san francisco")
}
