% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_script.R
\name{eq_map}
\alias{eq_map}
\title{Create an interactive map of historical earthquakes}
\usage{
eq_map(data, annot_col = NULL)
}
\arguments{
\item{data}{A dataframe with earthquake location columns \code{LATITUDE} and \code{LONGITUDE},
and magnatude column \code{EQ_PRIMARY}.  The included EQ_NOAA dataframe is already
correctly formateed, and is intended to be used with this function}

\item{annot_col}{The character name of the column in \code{data} containing a character
vector of optional popup text to be shown when the earthquake is clicked.  Defaults to
NULL, where no label is shown}
}
\value{
An html map object, to which further Leaflet objects can be added
}
\description{
Create an interactive map of historical earthquakes using the NOAA earthquake
database included with this package.  Earthquakes are plotted as circles with their
radii proportional to the magnatude of the earthquakes. Optionally, labels can be passed
which will pop up when the earthquake is clicked on the map
}
\details{
This function uses of the popular \code{leaflet} package, which creates interactive
html maps within R
}
\examples{
\dontrun{
data('clean_NOAA')
clean_NOAA \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  eq_map(annot_col = "DATE")
data('clean_NOAA')
clean_NOAA \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "DATE")
}

}
\references{
NOAA earthquake database: \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\seealso{
\pkg{leaflet}
}
