% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_script.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{A geom for labeling the timeline plot from \code{\link{geom_timeline}}}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "timeline_label",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, x_min = NULL, x_max = NULL, top_x_mag = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{x_min}{(optional) A Date object of the earliest data to be plotted}

\item{x_max}{(optional) A Date object of the latest data to be plotted}

\item{top_x_mag}{The top \code{x} labels to plot.  For example, if earthquake
occurrences are being plotted, and you only want to label the top 5, pass some
measure of magnatude in the aes slot and pass 5 here}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
This geom labels the timeline created from \code{\link{geom_timeline}}.  It draws
vertical lines up from the points on the timeline, with labels for the particular event.
You have the option of only labeleing the top \code{x} number of events sorted by a given
amount
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics (required are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{label}} The labels to be added
  \item \code{y}
  \item \code{magnatude} The feature whose value determines the top \code{x} to be labeled
  \item \code{color}
  \item \code{fill}
  \item \code{size}
  \item \code{alpha}
}
}

\examples{
data('clean_NOAA')
NOAA <- clean_NOAA
plot_NOAA <- NOAA[NOAA$COUNTRY=="USA" | NOAA$COUNTRY=="CANADA",]
xmin <- lubridate::ymd("2010-01-01")
xmax <- lubridate::ymd("2017-07-01")
plt <- ggplot2::ggplot(data = plot_NOAA, ggplot2::aes(x = DATE, y = COUNTRY)) +
  geom_timeline(ggplot2::aes(size = EQ_PRIMARY), x_min = xmin, x_max = xmax) +
  geom_timeline_label(ggplot2::aes(label = LOCATION_NAME, magnatude = EQ_PRIMARY),
     x_min = xmin, x_max = xmax, top_x_mag = 5) +
  ggplot2::theme_minimal()
print(plt)

}
