% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_script.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Create the ggproto object for geom_timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
This function creates the ggproto Geom object to plot the necessary grid grob for
\code{geom_timeline}
}
\examples{
#' NOAA <- data('clean_NOAA')
plot_NOAA <- NOAA[NOAA$COUNTRY=="USA" | NOAA$COUNTRY=="CANADA",]
xmin <- lubridate::ymd("2010-01-01")
xmax <- lubridate::ymd("2017-07-01")
plt <- ggplot2::ggplot(data = plot_NOAA, ggplot2::aes(x = DATE, y = COUNTRY)) +
  geom_timeline(ggplot2::aes(size = EQ_PRIMARY), x_min = xmin, x_max = xmax) +
  ggplot2::theme_minimal()
print(plt)

}
\keyword{datasets}
