% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_script.R
\name{my_draw_key_circle}
\alias{my_draw_key_circle}
\title{Allow geom_timeline to draw a useful legend}
\usage{
my_draw_key_circle(data, params, size)
}
\description{
This function is completely internal to geom_timeline, and
is not of use elsewhere
}
\examples{
# no examples appropriate, this function is completely internal to geom_timeline
}
