% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_script.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Create the ggproto object for stat_timeline}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimeline
}
\description{
This function creates the ggproto Stat object transform the data to effectively
plot the necessary grid grob for \code{stat_timeline}
}
\examples{
#' NOAA <- data('clean_NOAA')
plot_NOAA <- NOAA[NOAA$COUNTRY=="USA" | NOAA$COUNTRY=="CANADA",]
xmin <- lubridate::ymd("2010-01-01")
xmax <- lubridate::ymd("2017-07-01")
plt <- ggplot2::ggplot(data = plot_NOAA, ggplot2::aes(x = DATE, y = COUNTRY)) +
  geom_timeline(ggplot2::aes(size = EQ_PRIMARY), x_min = xmin, x_max = xmax) +
  ggplot2::theme_minimal()
print(plt)

}
\keyword{datasets}
