% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{title_case}
\alias{title_case}
\title{Convert a string to Title Case}
\usage{
title_case(char_vect)
}
\arguments{
\item{char_vect}{A character vector of the string to be converted to title case}
}
\value{
A character vector that is now Title Case
}
\description{
Have 'a string like this'? This function turns it into 'A String Like This'
}
\examples{
QuakeCoursera:::title_case('this is not title case')
QuakeCoursera:::title_case('THIS IS NOT TITLE CASE EITHER')
}
